fix bugs on schema field display
git-svn-id: svn://svn.cy55.de/Zope3/src/cybertools/trunk@2948 fd906abe-77d9-0310-91a1-e0d9ade77398

+++ media/tests.py
@@ -18,10 +18,12 @@ dataDir = os.path.join(os.path.dirname(media.__file__), 'testdata')
 def clearDataDir():
     for fn in os.listdir(dataDir):
         path = os.path.join(dataDir, fn)
-        if os.path.isdir(path):
-            for subfn in os.listdir(path):
-                os.unlink(os.path.join(path, subfn))
-            os.rmdir(path)
+        if not fn.startswith('.'):
+            if os.path.isdir(path):
+                for subfn in os.listdir(path):
+                    if not subfn.startswith('.'):
+                        os.unlink(os.path.join(path, subfn))
+                os.rmdir(path)
 
 
 class Test(unittest.TestCase):
